misc: - introduced CollectionUtils for various implicit enhancements, currently implemented a number of map key/value/key-value methods for Maps; - split parsing scopes into separate files for easier navigation; patch: - fixed a bug where strings surrounded with quotation marks and containing whitespaces (or other separator chars) were parsed incorrectly;
