+++ utils/src/main/scala/com/lomicron/utils/parsing/scopes/ParsingException.scala
@@ -1,0 +1,5 @@
+package com.lomicron.utils.parsing.scopes
+
+case class ParsingException(private val message: String = "",
+                            private val cause: Throwable = None.orNull)
+  extends RuntimeException(message, cause)
